+++ bundle/src/test/java/com/adobe/acs/commons/reports/models/ReferencesModelTest.java
@@ -19,11 +19,15 @@
  */
 package com.adobe.acs.commons.reports.models;
 
-import static org.junit.Assert.*;
-import static org.mockito.Mockito.when;
-
+import com.adobe.acs.commons.reports.internal.DelimiterConfiguration;
+import com.adobe.acs.commons.reports.models.references.MockReferenceList;
+import com.adobe.acs.commons.reports.models.references.MockReferencesAggregator;
+import com.adobe.granite.references.Reference;
+import com.adobe.granite.references.ReferenceAggregator;
+import com.adobe.granite.references.ReferenceList;
 import java.lang.reflect.Field;
-
+import java.util.HashMap;
+import java.util.Map;
 import org.apache.sling.api.resource.Resource;
 import org.junit.Before;
 import org.junit.Test;
@@ -32,15 +36,22 @@ import org.mockito.MockitoAnnotations;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.adobe.acs.commons.reports.models.references.MockReferenceList;
-import com.adobe.acs.commons.reports.models.references.MockReferencesAggregator;
-import com.adobe.granite.references.Reference;
-import com.adobe.granite.references.ReferenceAggregator;
-import com.adobe.granite.references.ReferenceList;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
+import static org.mockito.Mockito.when;
 
 public class ReferencesModelTest {
 
   private static final Logger log = LoggerFactory.getLogger(ReferencesModelTest.class);
+
+  private static final Map<String, String> CONFIG = new HashMap<String, String>() {{
+    put("field.delimiter", ",");
+    put("multi.value.delimiter", ";");
+  }};
+
+  @Mock
+  private DelimiterConfiguration delimiterConfiguration;
 
   @Mock
   private Resource mockResourceSource;
@@ -60,13 +71,15 @@ public class ReferencesModelTest {
     when(mockResourceSource.getPath()).thenReturn("/source");
     when(mockResourceTarget.getPath()).thenReturn("/target");
 
+    when(delimiterConfiguration.getMultiValueDelimiter()).thenReturn(CONFIG.get("multi.value.delimiter"));
+
     ReferenceList references = new MockReferenceList(mockResourceSource);
     references.add(new Reference(mockResourceSource, mockResourceTarget, "VALID"));
     references.add(new Reference(mockResourceSource, mockResourceSource, "INVALID"));
 
     ReferenceAggregator aggregator = new MockReferencesAggregator(references);
 
-    referencesModel = new ReferencesModel(mockResourceSource);
+    referencesModel = new ReferencesModel(mockResourceSource, delimiterConfiguration);
     Field af = ReferencesModel.class.getDeclaredField("aggregator");
     af.setAccessible(true);
     af.set(referencesModel, aggregator);
@@ -78,7 +91,7 @@ public class ReferencesModelTest {
     log.info("testGetReferences");
 
     assertNotNull(referencesModel.getReferences());
-    assertTrue(1 == referencesModel.getReferences().size());
+    assertEquals(1, referencesModel.getReferences().size());
     assertEquals("/target", referencesModel.getReferences().get(0).getTarget().getPath());
     assertEquals("VALID", referencesModel.getReferences().get(0).getType());
 
@@ -93,5 +106,24 @@ public class ReferencesModelTest {
     assertEquals("VALID - /target", referencesModel.getValue(mockResourceSource));
 
     log.info("Test Successful!");
+  }
+
+  @Test
+  public void testGetValueDelimiter() throws Exception {
+    ReferenceList references = new MockReferenceList(mockResourceSource);
+    references.add(new Reference(mockResourceSource, mockResourceTarget, "VALID"));
+    references.add(new Reference(mockResourceSource, mockResourceTarget, "VALID2"));
+
+    ReferenceAggregator aggregator = new MockReferencesAggregator(references);
+
+    referencesModel = new ReferencesModel(mockResourceSource, delimiterConfiguration);
+    Field af = ReferencesModel.class.getDeclaredField("aggregator");
+    af.setAccessible(true);
+    af.set(referencesModel, aggregator);
+    referencesModel.init();
+
+    String delimitedValue = referencesModel.getValue(mockResourceSource);
+    assertNotNull(delimitedValue);
+    assertTrue(delimitedValue.contains(delimiterConfiguration.getMultiValueDelimiter()));
   }
 }